[SPOT-292][REFACTOR] 게시글 및 댓글 컨트롤러 클래스 정리
